+++ src/integTest/groovy/com/monits/gradle/sca/FindbugsIntegTest.groovy
@@ -57,7 +57,7 @@ class FindbugsIntegTest extends AbstractPerSourceSetPluginIntegTestFixture {
         reportFile().assertContents(containsString('<Plugin id="jp.co.worksap.oss.findbugs" enabled="true"/>'))
 
         where:
-        version << ['2.3', '2.4', '2.7', '2.10', GradleVersion.current().version]
+        version << TESTED_GRADLE_VERSIONS
         findbugsVersion = ToolVersions.findbugsVersion
     }
 
@@ -168,7 +168,7 @@ class FindbugsIntegTest extends AbstractPerSourceSetPluginIntegTestFixture {
     @SuppressWarnings('MethodName')
     void 'Android generated classes are available'() {
         given:
-        writeAndroidBuildFile()
+        writeAndroidBuildFile(DEFAULT_ANDROID_VERSION)
         writeAndroidManifest()
         useEmptySuppressionFilter()
         file('src/main/res/values/strings.xml') <<
@@ -191,6 +191,7 @@ class FindbugsIntegTest extends AbstractPerSourceSetPluginIntegTestFixture {
 
         when:
         BuildResult result = gradleRunner()
+                .withGradleVersion(gradleVersionForAndroid(DEFAULT_ANDROID_VERSION))
                 .build()
 
         then:
@@ -235,9 +236,8 @@ class FindbugsIntegTest extends AbstractPerSourceSetPluginIntegTestFixture {
         reportFile().assertContents(containsString('<Errors errors="0" missingClasses="0">'))
 
         where:
-        // mockable android jar is available since 1.1.0
-        androidVersion << ['1.1.3', '1.2.3', '1.3.1', '1.5.0', '2.0.0', '2.1.0']
-        gradleVersion = androidVersion < '1.5.0' ? '2.9' : GradleVersion.current().version
+        androidVersion << AndroidLintIntegTest.ANDROID_PLUGIN_VERSIONS
+        gradleVersion = gradleVersionForAndroid(androidVersion)
     }
 
     @SuppressWarnings('MethodName')
@@ -247,6 +247,7 @@ class FindbugsIntegTest extends AbstractPerSourceSetPluginIntegTestFixture {
 
         when:
         BuildResult result = gradleRunner()
+                .withGradleVersion(gradleVersionForAndroid(DEFAULT_ANDROID_VERSION))
                 .build()
 
         then:
@@ -304,13 +305,14 @@ class FindbugsIntegTest extends AbstractPerSourceSetPluginIntegTestFixture {
     @SuppressWarnings('MethodName')
     void 'reports include just classes from their sourcesets'() {
         given:
-        writeAndroidBuildFile()
+        writeAndroidBuildFile(DEFAULT_ANDROID_VERSION)
         writeAndroidManifest()
         useEmptySuppressionFilter()
         goodCode()
 
         when:
         BuildResult result = gradleRunner()
+                .withGradleVersion(gradleVersionForAndroid(DEFAULT_ANDROID_VERSION))
                 .build()
 
         then: